Refactor byzantine fault files, propagate service exception back to caller
