+++ src/main/java/cloud/filibuster/junit/configuration/FilibusterAnalysisConfiguration.java
@@ -1,7 +1,7 @@
 package cloud.filibuster.junit.configuration;
 
 import cloud.filibuster.instrumentation.datatypes.Pair;
-import cloud.filibuster.junit.configuration.examples.byzantine.decoders.ByzantineDecoder;
+import cloud.filibuster.junit.configuration.examples.redis.byzantine.decoders.ByzantineDecoder;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import org.json.JSONObject;
