--- src/main/java/cloud/filibuster/junit/configuration/examples/byzantine/decoders/ByzantineDecoder.java
@@ -1,5 +1,0 @@
-package cloud.filibuster.junit.configuration.examples.byzantine.decoders;
-
-public enum ByzantineDecoder {
-    BYTE_ARRAY, STRING
-}
