+++ src/main/java/cloud/filibuster/junit/configuration/examples/redis/RedisSingleFaultCommandTimeoutExceptionAnalysisConfigurationFile.java
@@ -1,21 +1,21 @@
-package cloud.filibuster.junit.configuration.examples;
+package cloud.filibuster.junit.configuration.examples.redis;
 
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfiguration;
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfigurationFile;
 import cloud.filibuster.junit.configuration.FilibusterCustomAnalysisConfigurationFile;
-import cloud.filibuster.junit.configuration.examples.byzantine.decoders.ByzantineDecoder;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import static cloud.filibuster.instrumentation.Constants.REDIS_MODULE_NAME;
 
-public class RedisSingleGetStringByzantineFaultAnalysisConfigurationFile implements FilibusterAnalysisConfigurationFile {
+public class RedisSingleFaultCommandTimeoutExceptionAnalysisConfigurationFile implements FilibusterAnalysisConfigurationFile {
     private static final FilibusterCustomAnalysisConfigurationFile filibusterCustomAnalysisConfigurationFile;
 
-    private static <T> Map<String, T> createBzyantineFaultMap(T value) {
-        Map<String, T> myMap = new HashMap<>();
-        myMap.put("value", value);
+    private static Map<String, String> createErrorMap() {
+        Map<String, String> myMap = new HashMap<>();
+        myMap.put("cause", "Command timed out after 100 millisecond(s)");
+        myMap.put("code", "");
         return myMap;
     }
 
@@ -23,15 +23,10 @@ public class RedisSingleGetStringByzantineFaultAnalysisConfigurationFile impleme
         FilibusterCustomAnalysisConfigurationFile.Builder filibusterCustomAnalysisConfigurationFileBuilder = new FilibusterCustomAnalysisConfigurationFile.Builder();
 
         FilibusterAnalysisConfiguration.Builder filibusterAnalysisConfigurationBuilderRedisExceptions = new FilibusterAnalysisConfiguration.Builder()
-                .name("my_string_get_byzantine_fault")
-                .pattern(REDIS_MODULE_NAME + "/(get)\\b");
+                .name("io.lettuce.core.RedisCommandTimeoutException")
+                .pattern(REDIS_MODULE_NAME + "/(get|set)\\b");
 
-
-        // Potentially use junit-quickcheck to generate the possible values -> Would make the tests more "flaky"
-        String[] possibleValues = {null, "123", "", "abcd", "-123ABC", "ThisIsATestString"};
-        for (String value : possibleValues) {
-            filibusterAnalysisConfigurationBuilderRedisExceptions.byzantine("my_string_get_byzantine_fault", createBzyantineFaultMap(value), ByzantineDecoder.STRING);
-        }
+        filibusterAnalysisConfigurationBuilderRedisExceptions.exception("io.lettuce.core.RedisCommandTimeoutException", createErrorMap());
 
         filibusterCustomAnalysisConfigurationFileBuilder.analysisConfiguration(filibusterAnalysisConfigurationBuilderRedisExceptions.build());
