+++ src/main/java/cloud/filibuster/junit/configuration/examples/redis/byzantine/decoders/ByzantineDecoder.java
@@ -1,0 +1,5 @@
+package cloud.filibuster.junit.configuration.examples.redis.byzantine.decoders;
+
+public enum ByzantineDecoder {
+    BYTE_ARRAY, STRING
+}
